+++ CHANGELOG.md
@@ -25,6 +25,7 @@
 - `tlua::LuaRead`, `tlua::Push` & `tlua::PushInto` derive macros now support
     new-type style tuple structs: they are treated as the inner type.
 - `impl tlua::PushInto for Tuple`.
+- `net_box::promise::TryGet::into_res` and `From<TryGet<_, _>> for Result<_, _>`.
 - Doc-comments here and there.
 
 ### Fixed